pyproject.toml
[project]
name = "paid-social-nav"
version = "0.1.0"
description = "Unified social media advertising API for Meta, Reddit, Pinterest, TikTok, and X"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "google-cloud-bigquery>=3.11",
  "requests>=2.31",
  "python-json-logger>=2.0.7",
  "jinja2>=3.1.0",
  "anthropic>=0.34.0",
  "matplotlib>=3.7.0",
  "numpy>=1.24.0",
  "weasyprint>=60.0",
  "google-cloud-storage>=2.0.0",
  "google-auth>=2.0.0",
  "google-api-python-client>=2.0.0",
  "fastmcp>=2.13.1",
]

[project.scripts]
psn = "paid_social_nav.cli.main:app"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.27.0",
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["paid_social_nav*", "mcp_server*"]
exclude = ["tests*", "sql*", "infra*", "configs*"]

[tool.setuptools.package-data]
"paid_social_nav.render" = ["templates/*.j2"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = ["E501", "I001", "N812"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["paid_social_nav"]
omit = ["*/tests/*", "*/test_*"]
